+++ src/command_modules/azure-cli-cosmosdb/azure/cli/command_modules/cosmosdb/_help.py
@@ -23,17 +23,17 @@ helps['cosmosdb collection'] = """
 
 helps['cosmosdb check-name-exists'] = """
     type: command
-    short-summary: Checks that the Azure Cosmos DB account name already exists.
+    short-summary: Checks if an Azure Cosmos DB account name exists.
 """
 
 helps['cosmosdb create'] = """
     type: command
-    short-summary: Create a new Azure Cosmos DB database account.
+    short-summary: Creates a new Azure Cosmos DB database account.
 """
 
 helps['cosmosdb delete'] = """
     type: command
-    short-summary: Deletes an existing Azure Cosmos DB database account.
+    short-summary: Deletes an Azure Cosmos DB database account.
 """
 
 helps['cosmosdb failover-priority-change'] = """
@@ -43,35 +43,35 @@ helps['cosmosdb failover-priority-change'] = """
 
 helps['cosmosdb list'] = """
     type: command
-    short-summary: Lists all Azure Cosmos DB database accounts within a given resource group or subscription.
+    short-summary: List Azure Cosmos DB database accounts.
 """
 
 helps['cosmosdb list-connection-strings'] = """
     type: command
-    short-summary: Lists the connection strings for the specified Azure Cosmos DB database account.
+    short-summary: List the connection strings for a Azure Cosmos DB database account.
 """
 
 helps['cosmosdb list-keys'] = """
     type: command
-    short-summary: Lists the access keys for the specified Azure Cosmos DB database account.
+    short-summary: List the access keys for a Azure Cosmos DB database account.
 """
 
 helps['cosmosdb list-read-only-keys'] = """
     type: command
-    short-summary: Lists the read-only access keys for the specified Azure Cosmos DB database account.
+    short-summary: List the read-only access keys for a Azure Cosmos DB database account.
 """
 
 helps['cosmosdb regenerate-key'] = """
     type: command
-    short-summary: Regenerates an access key for the specified Azure Cosmos DB database account.
+    short-summary: Regenerate an access key for a Azure Cosmos DB database account.
 """
 
 helps['cosmosdb show'] = """
     type: command
-    short-summary: Retrieves the properties of an existing Azure Cosmos DB database account.
+    short-summary: Get the details of an Azure Cosmos DB database account.
 """
 
 helps['cosmosdb update'] = """
     type: command
-    short-summary: Update an existing Azure Cosmos DB database account.
+    short-summary: Update an Azure Cosmos DB database account.
 """