+++ packaged_releases/HISTORY.md
@@ -1,6 +1,63 @@
 Release History
 ===============
 
+2.0.17 (2017-09-11)
+-------------------
+
+* See the `HISTORY.rst` files in each module for a changelog.  
+
+Source download:
+https://azurecliprod.blob.core.windows.net/releases/azure-cli_packaged_2.0.17.tar.gz
+e0bb182cb53bdef8042cda1dd0e0e893b63aca5e415b053af2962e7a416eb072
+
+<details>
+  <summary>List of component versions in this release (click to expand)</summary><p>
+
+$ az --version
+
+azure-cli (2.0.17)  
+
+acr (2.0.11)  
+acs (2.0.15)  
+appservice (0.1.16)  
+batch (3.1.3)  
+billing (0.1.4)  
+cdn (0.0.8)  
+cloud (2.0.7)  
+cognitiveservices (0.1.7)  
+command-modules-nspkg (2.0.1)  
+component (2.0.7)  
+configure (2.0.10)  
+consumption (0.1.4)  
+container (0.1.10)  
+core (2.0.16)  
+cosmosdb (0.1.12)  
+dla (0.0.11)  
+dls (0.0.13)  
+eventgrid (0.1.3)  
+extension (0.0.2)  
+feedback (2.0.6)  
+find (0.2.6)  
+interactive (0.3.9)  
+iot (0.1.11)  
+keyvault (2.0.10)  
+lab (0.0.10)  
+monitor (0.0.9)  
+network (2.0.14)  
+nspkg (3.0.1)  
+profile (2.0.12)  
+rdbms (0.0.6)  
+redis (0.2.8)  
+resource (2.0.14)  
+role (2.0.11)  
+servicefabric (0.0.3)  
+sf (1.0.8)  
+sql (2.0.11)  
+storage (2.0.15)  
+vm (2.0.14)  
+</p></details>
+
+
 2.0.16 (2017-08-31)
 -------------------