+++ src/command_modules/azure-cli-cdn/HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 
-unreleased
+0.0.8 (2017-09-11)
 ++++++++++++++++++
 * `cdn custom-domain create`: Fixed 'CustomDomain is not interable' bug.